refactor(azure-bicep-resource): Refactor schema usage
