+++ lib/modules/datasource/azure-bicep-resource/index.spec.ts
@@ -13,11 +13,11 @@ describe('modules/datasource/azure-bicep-resource/index', () => {
       .reply(
         200,
         codeBlock`
-      {
-        "Resources": {},
-        "Functions": {}
-      }
-      `
+          {
+            "Resources": {},
+            "Functions": {}
+          }
+        `
       );
 
     const azureBicepResourceDatasource = new AzureBicepResourceDatasource();
@@ -35,26 +35,26 @@ describe('modules/datasource/azure-bicep-resource/index', () => {
       .reply(
         200,
         codeBlock`
-        {
-          "Resources": {},
-          "Functions": {
-            "microsoft.billing/billingaccounts": {
-              "2019-10-01-preview": [
-                {
-                  "RelativePath": "billing/microsoft.billing/2019-10-01-preview/types.json",
-                  "Index": 307
-                }
-              ],
-              "2020-05-01": [
-                {
-                  "RelativePath": "billing/microsoft.billing/2020-05-01/types.json",
-                  "Index": 287
-                }
-              ]
+          {
+            "Resources": {},
+            "Functions": {
+              "microsoft.billing/billingaccounts": {
+                "2019-10-01-preview": [
+                  {
+                    "RelativePath": "billing/microsoft.billing/2019-10-01-preview/types.json",
+                    "Index": 307
+                  }
+                ],
+                "2020-05-01": [
+                  {
+                    "RelativePath": "billing/microsoft.billing/2020-05-01/types.json",
+                    "Index": 287
+                  }
+                ]
+              }
             }
           }
-        }
-      `
+        `
       );
 
     const azureBicepResourceDatasource = new AzureBicepResourceDatasource();
@@ -85,20 +85,20 @@ describe('modules/datasource/azure-bicep-resource/index', () => {
       .reply(
         200,
         codeBlock`
-        {
-          "Resources": {
-            "Microsoft.Storage/storageAccounts@2015-05-01-preview": {
-              "RelativePath": "storage/microsoft.storage/2015-05-01-preview/types.json",
-              "Index": 31
+          {
+            "Resources": {
+              "Microsoft.Storage/storageAccounts@2015-05-01-preview": {
+                "RelativePath": "storage/microsoft.storage/2015-05-01-preview/types.json",
+                "Index": 31
+              },
+              "Microsoft.Storage/storageAccounts@2018-02-01": {
+                "RelativePath": "storage/microsoft.storage/2018-02-01/types.json",
+                "Index": 85
+              }
             },
-            "Microsoft.Storage/storageAccounts@2018-02-01": {
-              "RelativePath": "storage/microsoft.storage/2018-02-01/types.json",
-              "Index": 85
-            }
-          },
-          "Functions": {}
-        }
-      `
+            "Functions": {}
+          }
+        `
       );
 
     const azureBicepResourceDatasource = new AzureBicepResourceDatasource();
